requirements.txt
finnhub-python>=2.4.0
python-dotenv>=1.0.0
praw>=7.7.0
nltk>=3.8.0
matplotlib>=3.7.0
pytz>=2023.3